pyproject.toml
[project]
name = "fast-web-auth"
version = "0.1.0"
description = "Basic FastAPI Web AuthN"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "jinja2>=3.1.6",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.0",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]

[tool.fastapi]
entrypoint = "main:app"
